[Refactoring]Define boot module
