+++ bootx64/src/exit.rs
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
+use common::boot as kernelboot;
 use common::constant::INIT_RSP;
 use common::size::{Byte, Size};
 use uefi::table::boot;
@@ -7,7 +8,7 @@ use x86_64::{PhysAddr, VirtAddr};
 
 pub fn bootx64<'a>(
     mem_map: &'a mut [boot::MemoryDescriptor],
-    boot_info: common::BootInfo,
+    boot_info: kernelboot::Info,
     entry_addr: VirtAddr,
     kernel_addr: PhysAddr,
     bytes_kernel: Size<Byte>,
@@ -31,7 +32,7 @@ fn disable_interruption() -> () {
     }
 }
 
-fn jump_to_kernel(boot_info: common::BootInfo, entry_addr: VirtAddr) -> ! {
+fn jump_to_kernel(boot_info: kernelboot::Info, entry_addr: VirtAddr) -> ! {
     boot_info.set();
 
     unsafe {
