+++ kernel/src/graphics.rs
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
- 
+
 pub mod font;
 
 #[macro_use]
@@ -7,6 +7,7 @@ pub mod screen;
 
 use crate::common;
 use crate::common::constant::VRAM_ADDR;
+use common::boot;
 use core::ptr;
 
 // Copy trait is needed for constructing MouseCursor struct
@@ -37,7 +38,7 @@ pub struct Vram {
 }
 
 impl Vram {
-    pub fn new_from_boot_info(boot_info: &common::BootInfo) -> Self {
+    pub fn new_from_boot_info(boot_info: &boot::Info) -> Self {
         let (x_len, y_len) = boot_info.vram().resolution();
 
         Self {
